+++ CHANGES.rst
@@ -10,6 +10,8 @@ Fixes:
 
 - be able to use multiple importcss_file_filter files
   [vangheem]
+- Fix issue where if existing querystring path value is ".::1",
+  after edit, the wrong value will be selected
 
 - Calculate z-index for modals dynamically to always be on top
   [vangheem]